modules/TempResponse/requirements.txt
azure-iot-device~=2.0.0
rpi.gpio